+++ addons/portal/controllers/__init__.py
@@ -1,6 +1,5 @@
 from . import mail
 from . import message_reaction
 from . import portal
-from . import portal_thread
 from . import thread
 from . import web
