+++ addons/portal_rating/controllers/__init__.py
@@ -1,4 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import portal_chatter
+from . import thread
 from . import portal_rating
