+++ addons/portal_rating/controllers/thread.py
@@ -1,8 +1,11 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from odoo.fields import Domain
-from odoo.addons.portal.controllers import portal_thread
+
+from odoo.addons.portal.controllers.thread import PortalThreadController
 
 
-class PortalChatter(portal_thread.PortalChatter):
+class PortalRatingThreadController(PortalThreadController):
     def _get_non_empty_message_domain(self):
         return super()._get_non_empty_message_domain() | Domain("rating_value", "!=", False)
